Conflicts:
	.gitignore
	Makefile
	boards.cfg
	common/Makefile
	common/lcd.c
	drivers/dma/Makefile
	drivers/i2c/Makefile
	drivers/input/Makefile
	drivers/mmc/Makefile
	drivers/pcmcia/Makefile
	drivers/power/battery/Makefile
	drivers/power/pmic/Makefile
	drivers/rtc/Makefile
	drivers/serial/Makefile
	drivers/serial/serial.c
	drivers/spi/Makefile
	drivers/usb/gadget/Makefile
	drivers/video/Makefile
	fs/Makefile
	lib/Makefile
	tools/Makefile
	tools/bmp_logo.c

+++ arch/sandbox/include/asm/types.h
@@ -1,23 +1,7 @@
 /*
  * Copyright (c) 2011 The Chromium OS Authors.
  *
- * See file CREDITS for list of people who contributed to this
- * project.
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License as
- * published by the Free Software Foundation; either version 2 of
- * the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.	 See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston,
- * MA 02111-1307 USA
+ * SPDX-License-Identifier:	GPL-2.0+
  */
 
 #ifndef __ASM_SANDBOX_TYPES_H
@@ -64,8 +48,8 @@ typedef unsigned long long u64;
 #define BITS_PER_LONG	CONFIG_SANDBOX_BITS_PER_LONG
 
 typedef unsigned long dma_addr_t;
-typedef unsigned long phys_addr_t;
-typedef unsigned long phys_size_t;
+typedef u32 phys_addr_t;
+typedef u32 phys_size_t;
 
 #endif /* __KERNEL__ */
